% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extract_Match_Data.R
\name{extractdata}
\alias{extractdata}
\title{Scrape ball by ball data}
\usage{
extractdata(matchIDs)
}
\arguments{
\item{matchIDs}{insert list of match id}
}
\value{
Ball by ball data frame
}
\description{
Scrape ball by ball data
}
\examples{
extractdata(521)
extractdata(c(521,533))
}
\author{
Kusal De Silva and Rajitha M. Silva
}
