% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICC_Match_ID.R
\name{iccmatchdata}
\alias{iccmatchdata}
\title{Scrape match id data from ICC website}
\usage{
iccmatchdata(MatchFormat, TeamGender)
}
\arguments{
\item{Format}{of the match}

\item{Gender}{of the team}
}
\value{
data frame with ICC match id data
}
\description{
Scrape match id data from ICC website
}
\examples{
iccmatchdata('ODI', 'Men')
iccmatchdata('ODI', 'Women')

}
